Fix storeBooking function
The function used "setDoc" to store new bookings to the
database, so it replaced the entire document for dates
that already had a booked time block. An if statement
was added to check if the document exists. If it does,
setDoc with the merge option is used. If it doesn't,
setDoc is used without the merge option to create a
new document.